moving source code into single subfoler. This way my IDEA can seperate source code files from test code files. first suggestion move defaults to interface
