--- Race.java
@@ -1,9 +1,0 @@
-package rekees;
-
-/**
- *
- * @author Terry
- */
-public enum Race {
-    HUMAN, DWARF, ELF;
-}
