+++ src/rekees/NewCharacterDefaults.java
@@ -1,0 +1,14 @@
+package rekees;
+
+/**
+ * Created by Robert on 2015-10-31.
+ */
+public interface NewCharacterDefaults {
+    int INITIAL_LEVEL = 1;
+    int DEFAULT_STAT_VALUE = 10;
+    int DEFAULT_HP = 8;
+    int MIN_HP_VALUE = -10;
+    int MIN_STAT_VALUE = 0;
+    Race DEFAULT_RACE = Race.HUMAN;
+    Profession DEFAULT_PROFESSION = Profession.WARRIOR;
+}
