+++ src/rekees/Race.java
@@ -1,0 +1,9 @@
+package rekees;
+
+/**
+ *
+ * @author Terry
+ */
+public enum Race {
+    HUMAN, DWARF, ELF;
+}
